# Vectorize BinaryEW and UnaryEW ops