--- a/cpp/benchmarks/CMakeLists.txt
+++ b/cpp/benchmarks/CMakeLists.txt
@@ -1,5 +1,6 @@
 open3d_ispc_add_executable(benchmarks)
 
+add_subdirectory(benchmark_utilities)
 add_subdirectory(core)
 add_subdirectory(geometry)
 add_subdirectory(io)