--- a/cpp/benchmarks/benchmark_utilities/CMakeLists.txt
+++ b/cpp/benchmarks/benchmark_utilities/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(benchmarks PRIVATE
+    Rand.cpp
+)