--- a/cpp/benchmarks/benchmark_utilities/Rand.h
+++ b/cpp/benchmarks/benchmark_utilities/Rand.h
@@ -1,0 +1,45 @@
+// ----------------------------------------------------------------------------
+// -                        Open3D: www.open3d.org                            -
+// ----------------------------------------------------------------------------
+// The MIT License (MIT)
+//
+// Copyright (c) 2018-2021 www.open3d.org
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+// IN THE SOFTWARE.
+// ----------------------------------------------------------------------------
+
+#pragma once
+
+#include "open3d/core/Dtype.h"
+#include "open3d/core/Scalar.h"
+#include "open3d/core/SizeVector.h"
+#include "open3d/core/Tensor.h"
+
+namespace open3d {
+namespace benchmarks {
+
+/// Returns a Tensor with random values within the range \p range .
+core::Tensor Rand(const core::SizeVector& shape,
+                  size_t seed,
+                  const std::pair<core::Scalar, core::Scalar>& range,
+                  core::Dtype dtype,
+                  const core::Device& device = core::Device("CPU:0"));
+
+}  // namespace benchmarks
+}  // namespace open3d