--- a/cpp/benchmarks/core/CMakeLists.txt
+++ b/cpp/benchmarks/core/CMakeLists.txt
@@ -1,8 +1,10 @@
 target_sources(benchmarks PRIVATE
+    BinaryEW.cpp
     HashMap.cpp
     MemoryManager.cpp
     ParallelFor.cpp
     Reduction.cpp
+    UnaryEW.cpp
     Zeros.cpp
 )