#### Highlights:

- Add vectorized ISPC code for all BinaryEW and UnaryEW ops.
- Add benchmarks for all BinaryEW and UnaryEW ops.
- Fix rounding issue in PTS unit test by moving from `0.3 * 255 = 76.5 -> 76 | 77` to `0.4 * 255 = 102 -> 102`.
- Fix rounding issue for `sqrt`, `sin`, `cos`, `exp` which were compared for strict binary representation equality.

#### Design decisions:

- Only test 100, 100k, 100m entries to keep benchmark time modarate.
- Duplicate enums for UnaryEW and BinaryEW ops to avoid coupling with internal kernel API.

#### Benchmarks (Intel i7-10750H):

<details>
<summary>Scalar Code Path</summary>

```
[Open3D INFO] CPUInfo: 6 cores, 12 threads.
[Open3D INFO] ISAInfo: DISABLED instruction set used in ISPC code.
2021-09-27T12:03:17+02:00
Running ./bin/benchmarks
Run on (12 X 4787.91 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 256 KiB (x6)
  L3 Unified 12288 KiB (x1)
Load Average: 4.38, 4.92, 4.88
--------------------------------------------------------------------------------------
Benchmark                                            Time             CPU   Iterations
--------------------------------------------------------------------------------------
BinaryEW/Add__CPU_Int8__100                      0.006 ms        0.006 ms        92294
BinaryEW/Add__CPU_Int8__100000                   0.066 ms        0.066 ms         9192
BinaryEW/Add__CPU_Int8__100000000                 67.0 ms         67.0 ms           10
BinaryEW/Add__CPU_UInt8__100                     0.006 ms        0.006 ms        99381
BinaryEW/Add__CPU_UInt8__100000                  0.071 ms        0.071 ms         9905
BinaryEW/Add__CPU_UInt8__100000000                72.8 ms         72.8 ms           10
BinaryEW/Add__CPU_Int16__100                     0.007 ms        0.007 ms        96892
BinaryEW/Add__CPU_Int16__100000                  0.068 ms        0.068 ms        10980
BinaryEW/Add__CPU_Int16__100000000                76.9 ms         76.9 ms            9
BinaryEW/Add__CPU_UInt16__100                    0.007 ms        0.007 ms        96740
BinaryEW/Add__CPU_UInt16__100000                 0.072 ms        0.072 ms        10262
BinaryEW/Add__CPU_UInt16__100000000               81.6 ms         81.6 ms            9
BinaryEW/Add__CPU_Int32__100                     0.007 ms        0.007 ms        92086
BinaryEW/Add__CPU_Int32__100000                  0.074 ms        0.074 ms        10199
BinaryEW/Add__CPU_Int32__100000000                 101 ms          101 ms            7
BinaryEW/Add__CPU_UInt32__100                    0.007 ms        0.007 ms       108235
BinaryEW/Add__CPU_UInt32__100000                 0.065 ms        0.065 ms        11607
BinaryEW/Add__CPU_UInt32__100000000               91.1 ms         91.1 ms            8
BinaryEW/Add__CPU_Int64__100                     0.006 ms        0.006 ms       107799
BinaryEW/Add__CPU_Int64__100000                  0.070 ms        0.070 ms         9472
BinaryEW/Add__CPU_Int64__100000000                 132 ms          132 ms            5
BinaryEW/Add__CPU_UInt64__100                    0.007 ms        0.007 ms       117461
BinaryEW/Add__CPU_UInt64__100000                 0.070 ms        0.070 ms        10617
BinaryEW/Add__CPU_UInt64__100000000                130 ms          130 ms            5
BinaryEW/Add__CPU_Float32__100                   0.006 ms        0.006 ms        92350
BinaryEW/Add__CPU_Float32__100000                0.075 ms        0.075 ms        10125
BinaryEW/Add__CPU_Float32__100000000               105 ms          105 ms            7
BinaryEW/Add__CPU_Float64__100                   0.006 ms        0.006 ms       108744
BinaryEW/Add__CPU_Float64__100000                0.072 ms        0.072 ms        10496
BinaryEW/Add__CPU_Float64__100000000               134 ms          134 ms            5
BinaryEW/Sub__CPU_Int8__100                      0.007 ms        0.007 ms       107953
BinaryEW/Sub__CPU_Int8__100000                   0.069 ms        0.069 ms        10955
BinaryEW/Sub__CPU_Int8__100000000                 69.1 ms         69.0 ms           10
BinaryEW/Sub__CPU_UInt8__100                     0.007 ms        0.007 ms        94607
BinaryEW/Sub__CPU_UInt8__100000                  0.068 ms        0.068 ms        11007
BinaryEW/Sub__CPU_UInt8__100000000                69.3 ms         69.3 ms           10
BinaryEW/Sub__CPU_Int16__100                     0.007 ms        0.007 ms        95217
BinaryEW/Sub__CPU_Int16__100000                  0.072 ms        0.072 ms        10412
BinaryEW/Sub__CPU_Int16__100000000                81.4 ms         81.4 ms            9
BinaryEW/Sub__CPU_UInt16__100                    0.007 ms        0.007 ms       101460
BinaryEW/Sub__CPU_UInt16__100000                 0.073 ms        0.073 ms         9823
BinaryEW/Sub__CPU_UInt16__100000000               82.5 ms         82.5 ms            9
BinaryEW/Sub__CPU_Int32__100                     0.006 ms        0.006 ms       102502
BinaryEW/Sub__CPU_Int32__100000                  0.070 ms        0.070 ms        10842
BinaryEW/Sub__CPU_Int32__100000000                95.6 ms         95.6 ms            8
BinaryEW/Sub__CPU_UInt32__100                    0.007 ms        0.007 ms       106047
BinaryEW/Sub__CPU_UInt32__100000                 0.070 ms        0.070 ms        10838
BinaryEW/Sub__CPU_UInt32__100000000               96.0 ms         96.0 ms            8
BinaryEW/Sub__CPU_Int64__100                     0.006 ms        0.006 ms       107575
BinaryEW/Sub__CPU_Int64__100000                  0.083 ms        0.083 ms        10120
BinaryEW/Sub__CPU_Int64__100000000                 136 ms          136 ms            5
BinaryEW/Sub__CPU_UInt64__100                    0.006 ms        0.006 ms       110458
BinaryEW/Sub__CPU_UInt64__100000                 0.071 ms        0.071 ms        10660
BinaryEW/Sub__CPU_UInt64__100000000                132 ms          132 ms            5
BinaryEW/Sub__CPU_Float32__100                   0.007 ms        0.007 ms       112282
BinaryEW/Sub__CPU_Float32__100000                0.074 ms        0.074 ms        10046
BinaryEW/Sub__CPU_Float32__100000000               100 ms          100 ms            7
BinaryEW/Sub__CPU_Float64__100                   0.006 ms        0.006 ms       105522
BinaryEW/Sub__CPU_Float64__100000                0.071 ms        0.071 ms        10396
BinaryEW/Sub__CPU_Float64__100000000               135 ms          135 ms            5
BinaryEW/Mul__CPU_Int8__100                      0.007 ms        0.007 ms       109185
BinaryEW/Mul__CPU_Int8__100000                   0.073 ms        0.072 ms        10307
BinaryEW/Mul__CPU_Int8__100000000                 74.0 ms         73.8 ms           10
BinaryEW/Mul__CPU_UInt8__100                     0.007 ms        0.007 ms        94845
BinaryEW/Mul__CPU_UInt8__100000                  0.073 ms        0.073 ms        10348
BinaryEW/Mul__CPU_UInt8__100000000                73.1 ms         73.1 ms           10
BinaryEW/Mul__CPU_Int16__100                     0.007 ms        0.007 ms        96085
BinaryEW/Mul__CPU_Int16__100000                  0.069 ms        0.069 ms        11019
BinaryEW/Mul__CPU_Int16__100000000                76.6 ms         76.6 ms            9
BinaryEW/Mul__CPU_UInt16__100                    0.007 ms        0.007 ms        98405
BinaryEW/Mul__CPU_UInt16__100000                 0.073 ms        0.073 ms        10332
BinaryEW/Mul__CPU_UInt16__100000000               84.9 ms         84.9 ms            9
BinaryEW/Mul__CPU_Int32__100                     0.007 ms        0.007 ms        93032
BinaryEW/Mul__CPU_Int32__100000                  0.073 ms        0.073 ms        10250
BinaryEW/Mul__CPU_Int32__100000000                99.3 ms         99.3 ms            7
BinaryEW/Mul__CPU_UInt32__100                    0.006 ms        0.006 ms       107682
BinaryEW/Mul__CPU_UInt32__100000                 0.069 ms        0.069 ms        10835
BinaryEW/Mul__CPU_UInt32__100000000               94.9 ms         94.9 ms            7
BinaryEW/Mul__CPU_Int64__100                     0.006 ms        0.006 ms       102459
BinaryEW/Mul__CPU_Int64__100000                  0.073 ms        0.073 ms        10616
BinaryEW/Mul__CPU_Int64__100000000                 132 ms          132 ms            5
BinaryEW/Mul__CPU_UInt64__100                    0.006 ms        0.006 ms       112372
BinaryEW/Mul__CPU_UInt64__100000                 0.071 ms        0.071 ms        10561
BinaryEW/Mul__CPU_UInt64__100000000                132 ms          132 ms            5
BinaryEW/Mul__CPU_Float32__100                   0.006 ms        0.006 ms       113887
BinaryEW/Mul__CPU_Float32__100000                0.069 ms        0.069 ms        10610
BinaryEW/Mul__CPU_Float32__100000000              97.1 ms         97.1 ms            7
BinaryEW/Mul__CPU_Float64__100                   0.007 ms        0.007 ms       104010
BinaryEW/Mul__CPU_Float64__100000                0.076 ms        0.076 ms         9848
BinaryEW/Mul__CPU_Float64__100000000               137 ms          137 ms            5
BinaryEW/Div__CPU_Int8__100                      0.007 ms        0.007 ms       110795
BinaryEW/Div__CPU_Int8__100000                   0.089 ms        0.089 ms         8424
BinaryEW/Div__CPU_Int8__100000000                 89.8 ms         89.8 ms            8
BinaryEW/Div__CPU_UInt8__100                     0.007 ms        0.007 ms        80041
BinaryEW/Div__CPU_UInt8__100000                  0.084 ms        0.084 ms         8788
BinaryEW/Div__CPU_UInt8__100000000                84.5 ms         84.5 ms            8
BinaryEW/Div__CPU_Int16__100                     0.007 ms        0.007 ms        97493
BinaryEW/Div__CPU_Int16__100000                  0.089 ms        0.089 ms         8488
BinaryEW/Div__CPU_Int16__100000000                96.7 ms         96.7 ms            7
BinaryEW/Div__CPU_UInt16__100                    0.007 ms        0.007 ms        90192
BinaryEW/Div__CPU_UInt16__100000                 0.089 ms        0.089 ms         7472
BinaryEW/Div__CPU_UInt16__100000000               98.6 ms         98.5 ms            8
BinaryEW/Div__CPU_Int32__100                     0.007 ms        0.007 ms        99742
BinaryEW/Div__CPU_Int32__100000                  0.090 ms        0.090 ms         7250
BinaryEW/Div__CPU_Int32__100000000                 116 ms          116 ms            6
BinaryEW/Div__CPU_UInt32__100                    0.007 ms        0.007 ms       105764
BinaryEW/Div__CPU_UInt32__100000                 0.089 ms        0.089 ms         7012
BinaryEW/Div__CPU_UInt32__100000000                117 ms          117 ms            6
BinaryEW/Div__CPU_Int64__100                     0.007 ms        0.007 ms       104170
BinaryEW/Div__CPU_Int64__100000                  0.190 ms        0.190 ms         3874
BinaryEW/Div__CPU_Int64__100000000                 245 ms          245 ms            3
BinaryEW/Div__CPU_UInt64__100                    0.007 ms        0.007 ms       112143
BinaryEW/Div__CPU_UInt64__100000                 0.172 ms        0.172 ms         4297
BinaryEW/Div__CPU_UInt64__100000000                226 ms          226 ms            3
BinaryEW/Div__CPU_Float32__100                   0.007 ms        0.007 ms       110295
BinaryEW/Div__CPU_Float32__100000                0.071 ms        0.071 ms        10498
BinaryEW/Div__CPU_Float32__100000000              96.9 ms         96.9 ms            8
BinaryEW/Div__CPU_Float64__100                   0.007 ms        0.007 ms       104560
BinaryEW/Div__CPU_Float64__100000                0.072 ms        0.072 ms        10287
BinaryEW/Div__CPU_Float64__100000000               136 ms          136 ms            5
BinaryEW/LogicalAnd__CPU_Bool__100               0.007 ms        0.007 ms       107405
BinaryEW/LogicalAnd__CPU_Bool__100000            0.068 ms        0.068 ms         9611
BinaryEW/LogicalAnd__CPU_Bool__100000000          69.0 ms         69.0 ms           11
BinaryEW/LogicalAnd__CPU_Int8__100               0.007 ms        0.007 ms        95025
BinaryEW/LogicalAnd__CPU_Int8__100000            0.073 ms        0.073 ms        10931
BinaryEW/LogicalAnd__CPU_Int8__100000000          68.8 ms         68.8 ms           10
BinaryEW/LogicalAnd__CPU_UInt8__100              0.007 ms        0.007 ms        83697
BinaryEW/LogicalAnd__CPU_UInt8__100000           0.068 ms        0.068 ms        10625
BinaryEW/LogicalAnd__CPU_UInt8__100000000         69.2 ms         69.2 ms           10
BinaryEW/LogicalAnd__CPU_Int16__100              0.007 ms        0.007 ms        94774
BinaryEW/LogicalAnd__CPU_Int16__100000           0.069 ms        0.069 ms         9615
BinaryEW/LogicalAnd__CPU_Int16__100000000         69.7 ms         69.7 ms           10
BinaryEW/LogicalAnd__CPU_UInt16__100             0.007 ms        0.007 ms        88828
BinaryEW/LogicalAnd__CPU_UInt16__100000          0.069 ms        0.069 ms        10592
BinaryEW/LogicalAnd__CPU_UInt16__100000000        69.8 ms         69.8 ms           10
BinaryEW/LogicalAnd__CPU_Int32__100              0.007 ms        0.007 ms        94347
BinaryEW/LogicalAnd__CPU_Int32__100000           0.069 ms        0.069 ms         9428
BinaryEW/LogicalAnd__CPU_Int32__100000000         70.7 ms         70.7 ms           10
BinaryEW/LogicalAnd__CPU_UInt32__100             0.007 ms        0.007 ms        94859
BinaryEW/LogicalAnd__CPU_UInt32__100000          0.069 ms        0.069 ms         9075
BinaryEW/LogicalAnd__CPU_UInt32__100000000        70.7 ms         70.7 ms           10
BinaryEW/LogicalAnd__CPU_Int64__100              0.007 ms        0.007 ms        97844
BinaryEW/LogicalAnd__CPU_Int64__100000           0.070 ms        0.070 ms        10694
BinaryEW/LogicalAnd__CPU_Int64__100000000         75.5 ms         75.5 ms           10
BinaryEW/LogicalAnd__CPU_UInt64__100             0.007 ms        0.007 ms       103679
BinaryEW/LogicalAnd__CPU_UInt64__100000          0.070 ms        0.070 ms        10785
BinaryEW/LogicalAnd__CPU_UInt64__100000000        76.9 ms         76.9 ms            9
BinaryEW/LogicalAnd__CPU_Float32__100            0.007 ms        0.007 ms       103210
BinaryEW/LogicalAnd__CPU_Float32__100000         0.084 ms        0.084 ms         8900
BinaryEW/LogicalAnd__CPU_Float32__100000000       80.8 ms         80.8 ms            9
BinaryEW/LogicalAnd__CPU_Float64__100            0.007 ms        0.007 ms        99561
BinaryEW/LogicalAnd__CPU_Float64__100000         0.084 ms        0.084 ms         8946
BinaryEW/LogicalAnd__CPU_Float64__100000000       86.8 ms         86.8 ms            8
BinaryEW/LogicalOr__CPU_Bool__100                0.007 ms        0.007 ms       102694
BinaryEW/LogicalOr__CPU_Bool__100000             0.073 ms        0.073 ms        10268
BinaryEW/LogicalOr__CPU_Bool__100000000           72.6 ms         72.6 ms           10
BinaryEW/LogicalOr__CPU_Int8__100                0.007 ms        0.007 ms        94184
BinaryEW/LogicalOr__CPU_Int8__100000             0.082 ms        0.082 ms         9259
BinaryEW/LogicalOr__CPU_Int8__100000000           80.2 ms         80.2 ms            9
BinaryEW/LogicalOr__CPU_UInt8__100               0.007 ms        0.007 ms        95689
BinaryEW/LogicalOr__CPU_UInt8__100000            0.081 ms        0.081 ms         9275
BinaryEW/LogicalOr__CPU_UInt8__100000000          78.8 ms         78.8 ms            9
BinaryEW/LogicalOr__CPU_Int16__100               0.007 ms        0.007 ms        96365
BinaryEW/LogicalOr__CPU_Int16__100000            0.081 ms        0.081 ms         8697
BinaryEW/LogicalOr__CPU_Int16__100000000          79.3 ms         79.3 ms            9
BinaryEW/LogicalOr__CPU_UInt16__100              0.007 ms        0.007 ms        96851
BinaryEW/LogicalOr__CPU_UInt16__100000           0.078 ms        0.078 ms         8447
BinaryEW/LogicalOr__CPU_UInt16__100000000         78.4 ms         78.4 ms            9
BinaryEW/LogicalOr__CPU_Int32__100               0.007 ms        0.007 ms        97841
BinaryEW/LogicalOr__CPU_Int32__100000            0.082 ms        0.082 ms         9203
BinaryEW/LogicalOr__CPU_Int32__100000000          78.0 ms         78.0 ms            9
BinaryEW/LogicalOr__CPU_UInt32__100              0.007 ms        0.007 ms        99632
BinaryEW/LogicalOr__CPU_UInt32__100000           0.079 ms        0.079 ms         9625
BinaryEW/LogicalOr__CPU_UInt32__100000000         80.1 ms         80.1 ms            9
BinaryEW/LogicalOr__CPU_Int64__100               0.007 ms        0.007 ms        96538
BinaryEW/LogicalOr__CPU_Int64__100000            0.083 ms        0.083 ms         9196
BinaryEW/LogicalOr__CPU_Int64__100000000          82.5 ms         82.5 ms            9
BinaryEW/LogicalOr__CPU_UInt64__100              0.007 ms        0.007 ms       104072
BinaryEW/LogicalOr__CPU_UInt64__100000           0.077 ms        0.077 ms         8422
BinaryEW/LogicalOr__CPU_UInt64__100000000         81.6 ms         81.6 ms            9
BinaryEW/LogicalOr__CPU_Float32__100             0.007 ms        0.007 ms       103610
BinaryEW/LogicalOr__CPU_Float32__100000          0.074 ms        0.074 ms        10081
BinaryEW/LogicalOr__CPU_Float32__100000000        76.5 ms         76.5 ms           10
BinaryEW/LogicalOr__CPU_Float64__100             0.007 ms        0.007 ms       101742
BinaryEW/LogicalOr__CPU_Float64__100000          0.082 ms        0.082 ms         9200
BinaryEW/LogicalOr__CPU_Float64__100000000        86.3 ms         86.3 ms            9
BinaryEW/LogicalXor__CPU_Bool__100               0.007 ms        0.007 ms       105057
BinaryEW/LogicalXor__CPU_Bool__100000            0.068 ms        0.068 ms        11001
BinaryEW/LogicalXor__CPU_Bool__100000000          68.7 ms         68.8 ms           11
BinaryEW/LogicalXor__CPU_Int8__100               0.007 ms        0.007 ms        95440
BinaryEW/LogicalXor__CPU_Int8__100000            0.064 ms        0.064 ms         9684
BinaryEW/LogicalXor__CPU_Int8__100000000          64.6 ms         64.6 ms           11
BinaryEW/LogicalXor__CPU_UInt8__100              0.007 ms        0.007 ms        94815
BinaryEW/LogicalXor__CPU_UInt8__100000           0.064 ms        0.064 ms        11391
BinaryEW/LogicalXor__CPU_UInt8__100000000         65.9 ms         65.9 ms           10
BinaryEW/LogicalXor__CPU_Int16__100              0.007 ms        0.007 ms        92350
BinaryEW/LogicalXor__CPU_Int16__100000           0.070 ms        0.070 ms        10850
BinaryEW/LogicalXor__CPU_Int16__100000000         69.3 ms         69.3 ms           10
BinaryEW/LogicalXor__CPU_UInt16__100             0.007 ms        0.007 ms        91966
BinaryEW/LogicalXor__CPU_UInt16__100000          0.074 ms        0.074 ms        10242
BinaryEW/LogicalXor__CPU_UInt16__100000000        74.3 ms         74.3 ms           10
BinaryEW/LogicalXor__CPU_Int32__100              0.007 ms        0.007 ms        95000
BinaryEW/LogicalXor__CPU_Int32__100000           0.065 ms        0.065 ms        11628
BinaryEW/LogicalXor__CPU_Int32__100000000         65.9 ms         65.9 ms           10
BinaryEW/LogicalXor__CPU_UInt32__100             0.007 ms        0.007 ms       101700
BinaryEW/LogicalXor__CPU_UInt32__100000          0.073 ms        0.073 ms        10208
BinaryEW/LogicalXor__CPU_UInt32__100000000        75.8 ms         75.8 ms            9
BinaryEW/LogicalXor__CPU_Int64__100              0.007 ms        0.007 ms        98333
BinaryEW/LogicalXor__CPU_Int64__100000           0.070 ms        0.070 ms        10752
BinaryEW/LogicalXor__CPU_Int64__100000000         75.9 ms         75.9 ms            9
BinaryEW/LogicalXor__CPU_UInt64__100             0.007 ms        0.007 ms       101294
BinaryEW/LogicalXor__CPU_UInt64__100000          0.074 ms        0.074 ms        10110
…